Move platform files to correct directory
