+++ src/components/platform/platformListDisplayer.tsx
@@ -3,6 +3,9 @@ import {usePlatformsFetch} from "../../clients/platformClient";
 import PlatformList from "./platformList";
 import {PageControl} from "../pagination/pageControl";
 import PlatformFilter from "../../models/platformFilter";
+import {Link} from "react-router-dom";
+import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
+import {faSquarePlus} from "@fortawesome/free-solid-svg-icons/faSquarePlus";
 
 export default function PlatformListDisplayer() {
     const [page, setPage] = useState(1)
@@ -12,6 +15,11 @@ export default function PlatformListDisplayer() {
     if (loading)
         return <div>Loading...</div>
     return <>
+        <div className='row mb-4'>
+            <div className='col-12 offset-sm-10 col-sm-2 d-flex'>
+                <Link to='/platforms/create' title='Create new platrofm' className='btn btn-primary flex-fill'><FontAwesomeIcon icon={faSquarePlus}/></Link>
+            </div>
+        </div>
         <div className='row mb-4'>
             <div className='col-12'>
                 <div className='row'>
